safari-600.7-branch/LayoutTests/editing/pasteboard/paste-with-redundant-style.html

<!DOCTYPE html>
<html>
<head>
<style>
em { text-decoration: underline; font-style: none; }
</style>
</head>
<body>
<div id="test" contenteditable><b style="font-style: italic;"><br></b></div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

document.getElementById('test').focus();
document.execCommand('insertHTML', false,
    '<em style="font-style: none; text-decoration: underline; font-weight: bold;"><span style="color: black;">hello world</span></em><br>'
    + '<span style="font-weight: bold;"><span class="Apple-style-span" style="font-weight: bold;">WebKit</span></span><br>'
    + '<em><span style="font-style: italic;">rocks</span></em>');

Markup.description('This tests pasting content with redundant inline style.\n'
    + 'You should not see any inline styles other than font-weight: bold and there should be exactly one span around WebKit.');
Markup.dump('test');

</script>
</body>
</html>
