Safari-609.1.11.1/LayoutTests/editing/pasteboard/paste-with-redundant-style-expected.txt

This tests pasting content with redundant inline style.
You should not see any inline styles other than font-weight: bold and there should be exactly one span around WebKit.
| <em>
|   style="font-weight: bold;"
|   "hello world"
| <br>
| <span>
|   style="font-weight: bold;"
|   "WebKit"
| <br>
| <em>
|   "rocks<#selection-caret>"
| <b>
|   style="font-style: italic;"
|   <br>
